ml/requirements.txt
torch
torch-geometric
scikit-learn
pandas
numpy
censusgeocode
psycopg2-binary
python-dotenv
tqdm
